sandbox: Add a test for disabling CONFIG_CMDLINE
Now that everything is working, add a test to make sure that this
builds correctly.

Series-to: u-boot
Series-cc: trini
Series-links: 376591 1:374629
Series-changes: 3
- Rebase on Tom's LONGHELP series
- Correct 'of' typo

Series-version: 3
Cover-letter:
Tidy up use of CONFIG_CMDLINE
It should be possible to disable CONFIG_CMDLINE and have all commands
and related functionality dropped from U-Boot. This is useful when
trying to reduce the size of U-Boot.

Recent changes have stopped this from working.

This series repairs the feature for sandbox and adds a test to stop it
breaking again.

Note that quite a lot of functionality is lost if CONFIG_CMDLINE is
disabled, e.g. networking and most booting options. Further work is
needed to make the option more generally useful.
END

Signed-off-by: Simon Glass <sjg@chromium.org>