bootstd: Add a menu option to bootflow scan
Allow showing a menu and automatically booting, with 'bootflow scan'.
This is more convenient than using a script.

Series-to: u-boot
Series-links: 370485
Series-version: 3

Cover-letter:
Resolve issues with booting distros on x86
This little series reprises the EFI-video fix, fixes a USB problem and
enables a boot script for coreboot.

It also provides a -m flag for 'bootflow scan' so it can display a menu
automatically.

With these changes it is possible to boot a Linux distro automatically
with U-Boot on x86, including when U-Boot is the second-stage
bootloader.
END

Signed-off-by: Simon Glass <sjg@chromium.org>