FROMLIST: x86: serial: ns16550: Allow the UART to be silently disabled
U-Boot normally requires a UART. When booting from coreboot it is
sometimes just not available, e.g. when no sysinfo or DBG2 information
is provided.

In this case we need to continue running, since the display can be used.
Add a flag to disable serial for this case.

This allows U-Boot to start up and operation from the display, instead
of hanging on start-up.

This could perhaps be hidden behind a Kconfig option to reduce code
size.

https://patchwork.ozlabs.org/project/uboot/patch/20231101180447.99361-1-sjg@chromium.org/

BUG=b:303689867
TEST=boot altfw on skyrim; see that U-Boot appears on the display

Change-Id: I7a5591deb3c3e2bb1eb688743886579ebffb2639
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/u-boot/+/5005746
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Commit-Queue: Karthikeyan Ramasubramanian <kramasub@google.com>