bootm: Drop arguments from do_bootm_states() and rename
Create a new struct which holds the information required by bootm. Set
this up for each existing caller.

Now that none of the functions called from do_bootm_states() need an
argv[] list, change the arguments of do_bootm_states() as well.

For booti make sure it only uses argv[] and argc at the top of the
function, so we can eventually refactor to remove these parameters.

Finally, rename the function to bootm_run_states() to better indicate
its purpose. The 'do_' prefix is used to indicate a command processor,
which this is now not.

Signed-off-by: Simon Glass <sjg@chromium.org>