image: Correct load_bug typo
Correct a typo in the function comment for image_decomp().

Series-to: u-boot
Series-cc: trini, marek
Cover-letter:
bootm: Handle compressed arm64 images with bootm
This little series corrects a problem I noticed with arm64 images,
where the kernel is not recognised:

   ## Loading kernel from FIT Image at 10000000 ...
      Using 'conf-930' configuration
      Verifying Hash Integrity ... OK
      Trying 'kernel' kernel subimage
        Description:  Linux-6.6.0-rc7-next-20231024-00003-g259c196f194c
        Type:         Kernel Image (no loading done)
        Compression:  gzip compressed
        Data Start:   0x10000138
        Data Size:    13667956 Bytes = 13 MiB
      Verifying Hash Integrity ... OK
   Bad Linux ARM64 Image magic!

The problem is that the arm64 magic is checked before the image is
decompressed.

Another issue is that the load address is read from the 'load' property
even with a kernel_noload image. This means that the kernel is loaded
to address 0, which may not be valid on the board. We can use the
kernel_addr_r environment variable instead.

A patch is included to show the kernel load-address, so it is easy to
see what is going on.
END

Change-Id: I33852d11a04a125d8a4fd7938a8eb989454e7fd3
Signed-off-by: Simon Glass <sjg@chromium.org>