bootm: Create a function to run through the booti states
In a few places, the booti command is used to handle a boot. We want
these to be done without needing CONFIG_CMDLINE, so add a new
booti_run() function to handle this.

So far this is not used.

Series-to: u-boot
Series-cc: trini
Series-version: 2
Series-links: 385007
Cover-letter:
Complete decoupling of bootm logic from commands
This series continues refactoring the bootm code to allow it to be used
with CONFIG_COMMAND disabled. The OS-handling code is refactored and
a new bootm_run() function is created to run through the bootm stages.
This completes the work.

A booti_go() function is created also, in case it proves useful, but at
last for now standard boot does not use this.

This is cmdd (part d of CMDLINE refactoring)
It depends on dm/bootstda-working
which depends on dm/cmdc-working
END

Series-changes: 2
- Update series to allow OS access to cmdline arguments for bootm

Change-Id: I95fe77f4a7f3d6c3dba3ecb3aec30a48739198f9
Signed-off-by: Simon Glass <sjg@chromium.org>