x86: zboot: Tidy up the comment for zboot_run()
The current use case (ChromeOS) is the uncommon case. Document how this
function is more normally used, where base is 0 and cmdline is NULL

Series-to: u-boot
Series-cc: trini
Cover-letter:
Complete decoupling of zboot logic from commands
This series refactors the zboot code to allow it to be used with
CONFIG_COMMAND disabled.

A new zboot_run() function is used to boot a zimage.
END

Signed-off-by: Simon Glass <sjg@chromium.org>