config: rockchip: Add a rockpro64 boot script for reven
Add a simple boot script to boot from NVMe since this is faster for
development than MMC.

Expand bootstage to cope with the number of records generated.

Also increase the bootdelay to 5 seconds, so there is more time to stop
the boot.

Init PCI so that NVMe works better. It still sometimes fails with an
error:

   rockchip_pcie pcie@f8000000: PCIe link training gen1 timeout!

BUG=b:311172947
TEST=with other CLS, boot into ChromeOS on rockpro64 board and see the
logo shown

Change-Id: I8b4cd035d1b29d4b027338d906185ba84fde1db5
Signed-off-by: Simon Glass <sjg@chromium.org>