bootflow: Allow adding an argument
When an existing argument is set to a new value, the normal behaviour
is to overwrite the value of the existing argument.

Sometimes it is useful to add the argument, so that there are two.
For example there can be multiple 'console=' arguments passed to
Linux.

Add support for this.

Change-Id: Ia13d4f826a90f0bb8782e30f5931279120bb90f3
Signed-off-by: Simon Glass <sjg@chromium.org>