CHROMIUM: Support decompressing non-kernel OS images
Sometimes the kernel is built as an EFI application rather than a
binary. We still want to support compression for this case.

For arm64 the entry point is set later in the bootm_load_os() function,
since these images are typically relocated due to the 2MB-alignment
requirement of arm64 images. But since the EFI image is not in the same
format, we need to update the entry point earlier.

Set the entry point always, for kernel_noload to resolve this problem.
It should be harmless to do this always.

BUG=b:316629213
BRANCH=none
TEST=build and boot EFI kernel (see bug)

Change-Id: If0fab868c705be493705cf2609f17e3c85b1cefa